Separated styling into their own files
