+++ app/auth.tsx
@@ -1,5 +1,5 @@
 import AuthScreen from "../src/screens/AuthScreen";
 
-export default function Index() {
+export default function AuthPage() {
   return <AuthScreen />;
 }
