+++ app/login.tsx
@@ -1,70 +1,5 @@
-import React, { useState } from "react";
-import { View, StyleSheet } from "react-native";
-import { Text, Button, TextInput, useTheme } from "react-native-paper";
-import { router } from "expo-router";
+import LoginScreen from "@/src/screens/LoginScreen";
 
-export default function LoginScreen() {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const theme = useTheme();
-
-  const handleLogin = () => {
-    // Simulate login success
-    console.log("Logging in...", email);
-    router.replace("./GroceryList");
-  };
-
-  return (
-    <View
-      style={[styles.container, { backgroundColor: theme.colors.background }]}
-    >
-      <Text variant="headlineMedium" style={styles.title}>
-        Welcome Back
-      </Text>
-
-      <TextInput
-        label="Email"
-        value={email}
-        onChangeText={setEmail}
-        style={styles.input}
-        mode="outlined"
-      />
-      <TextInput
-        label="Password"
-        value={password}
-        onChangeText={setPassword}
-        secureTextEntry
-        style={styles.input}
-        mode="outlined"
-      />
-
-      <Button mode="contained" onPress={handleLogin} style={styles.button}>
-        Log In
-      </Button>
-
-      <Button onPress={() => router.push("./register")} style={styles.link}>
-        Don't have an account? Sign up
-      </Button>
-    </View>
-  );
+export default function AuthPage() {
+  return <LoginScreen />;
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    padding: 24,
-    justifyContent: "center",
-  },
-  title: {
-    marginBottom: 32,
-  },
-  input: {
-    marginBottom: 16,
-  },
-  button: {
-    marginTop: 16,
-  },
-  link: {
-    marginTop: 16,
-  },
-});
