+++ app/register.tsx
@@ -1,82 +1,5 @@
-import React, { useState } from "react";
-import { View, StyleSheet } from "react-native";
-import { Text, Button, TextInput, useTheme } from "react-native-paper";
-import { router } from "expo-router";
+import RegisterScreen from "@/src/screens/RegisterScreen";
 
-export default function RegisterScreen() {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [confirmPassword, setConfirmPassword] = useState("");
-  const theme = useTheme();
-
-  const handleRegister = () => {
-    if (password !== confirmPassword) {
-      alert("Passwords do not match");
-      return;
-    }
-    console.log("Registering...", email);
-    router.replace("./GroceryList");
-  };
-
-  return (
-    <View
-      style={[styles.container, { backgroundColor: theme.colors.background }]}
-    >
-      <Text variant="headlineMedium" style={styles.title}>
-        Create Account
-      </Text>
-
-      <TextInput
-        label="Email"
-        value={email}
-        onChangeText={setEmail}
-        style={styles.input}
-        mode="outlined"
-      />
-      <TextInput
-        label="Password"
-        value={password}
-        onChangeText={setPassword}
-        secureTextEntry
-        style={styles.input}
-        mode="outlined"
-      />
-      <TextInput
-        label="Confirm Password"
-        value={confirmPassword}
-        onChangeText={setConfirmPassword}
-        secureTextEntry
-        style={styles.input}
-        mode="outlined"
-      />
-
-      <Button mode="contained" onPress={handleRegister} style={styles.button}>
-        Sign Up
-      </Button>
-
-      <Button onPress={() => router.push("/login")} style={styles.link}>
-        Already have an account? Log in
-      </Button>
-    </View>
-  );
+export default function RegisterPage() {
+  return <RegisterScreen />;
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    padding: 24,
-    justifyContent: "center",
-  },
-  title: {
-    marginBottom: 32,
-  },
-  input: {
-    marginBottom: 16,
-  },
-  button: {
-    marginTop: 16,
-  },
-  link: {
-    marginTop: 16,
-  },
-});
