+++ src/screens/styles/AuthScreenStyles.ts
@@ -1,0 +1,25 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: "center",
+    padding: 24,
+    alignItems: "center",
+  },
+  title: {
+    marginBottom: 16,
+    textAlign: "center",
+  },
+  subtitle: {
+    marginBottom: 48,
+    textAlign: "center",
+    opacity: 0.7,
+  },
+  button: {
+    width: "100%",
+    marginBottom: 16,
+  },
+});
+
+export default styles;
