+++ src/screens/styles/GroceryListScreenStyles.ts
@@ -1,0 +1,24 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    padding: 20,
+    paddingTop: 60,
+  },
+  headline: {
+    marginBottom: 20,
+  },
+  input: {
+    marginBottom: 10,
+  },
+  addButton: {
+    marginBottom: 20,
+  },
+  emptyText: {
+    textAlign: "center",
+    color: "gray",
+  },
+});
+
+export default styles;
