+++ src/screens/styles/RegisterScreenStyles.ts
@@ -1,0 +1,23 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    padding: 24,
+    justifyContent: "center",
+  },
+  title: {
+    marginBottom: 32,
+  },
+  input: {
+    marginBottom: 16,
+  },
+  button: {
+    marginTop: 16,
+  },
+  link: {
+    marginTop: 16,
+  },
+});
+
+export default styles;
